+++ R/visualisation_recipe.R
@@ -3,9 +3,9 @@
 #' This function prepares objects for visualisation by returning a list of
 #' layers with data and geoms that can be easily plotted using for instance
 #' `ggplot2`. See the documentation for your object's class:
-#' \itemize{
-#'  \item{[modelbased][modelbased::visualisation_recipe.estimate_predicted] (`estimate_means`, `estimate_contrasts`, `estimate_slopes`, `estimate_predicted`, `estimate_grouplevel`)}
-#' }
+#'
+#'  - {[modelbased][modelbased::visualisation_recipe.estimate_predicted] (`estimate_means`, `estimate_contrasts`, `estimate_slopes`, `estimate_predicted`, `estimate_grouplevel`)}
+#'
 #'
 #' @param x An `easystats` object.
 #' @param ... Other arguments passed to other functions.